refactor(bzm2): extract hash thread work helpers
